CI: Refactor lfn_volume_info tests
